DHFPROD-9752: Convertion of mjs files under data hub 5 directory. (partial)
