+++ marklogic-data-hub/src/main/resources/ml-modules/root/data-hub/5/artifacts/stepDefinition.mjs
@@ -15,7 +15,7 @@
  */
 'use strict';
 
-import config from "/com.marklogic.hub/config.sjs";
+import config from "/com.marklogic.hub/config.mjs";
 import hubUtils from "/data-hub/5/impl/hub-utils.mjs";
 import StepDefinition from "/data-hub/5/impl/stepDefinition.mjs";
 const collections = ['http://marklogic.com/data-hub/step-definition'];
