+++ marklogic-data-hub/src/main/resources/ml-modules/root/data-hub/5/impl/entity-lib.mjs
@@ -21,14 +21,13 @@
  * resides in this module to promote reuse and also simplify upgrades as Entity Services changes within MarkLogic.
  */
 
-const sem = require("/MarkLogic/semantics.xqy");
+import sem from "/MarkLogic/semantics.xqy";
 const semPrefixes = {es: 'http://marklogic.com/entity-services#'};
-const config = require("/com.marklogic.hub/config.sjs");
-const consts = require("/data-hub/5/impl/consts.sjs");
-const hent = require("/data-hub/5/impl/hub-entities.xqy");
-const httpUtils = require("/data-hub/5/impl/http-utils.sjs");
-const hubUtils = require("/data-hub/5/impl/hub-utils.sjs");
-const {getModel} = require("../builtins/steps/mapping/default/lib.sjs");
+import config from "/com.marklogic.hub/config.mjs";
+import consts from "/data-hub/5/impl/consts.mjs";
+import hent from "/data-hub/5/impl/hub-entities.xqy";
+import httpUtils from "/data-hub/5/impl/http-utils.mjs";
+import hubUtils from "/data-hub/5/impl/hub-utils.mjs";
 
 /**
  * @return an array of strings, one for each EntityType
@@ -826,7 +825,7 @@ function findModelAndPropertyReferencesInMappingSteps(entityName, entityTypeId, 
   return cts.search(stepQuery).toArray().map(step => step.toObject().name);
 }
 
-module.exports = {
+export {
   deleteDraftModel,
   findForeignKeyReferencesInOtherModels,
   findModelReferencesInSteps,
