+++ marklogic-data-hub/src/main/resources/ml-modules/root/data-hub/5/impl/http-utils.mjs
@@ -47,7 +47,7 @@ function throwMethodNotSupported(message) {
   fn.error(null, ERROR_CODE, Sequence.from([405, message]));
 }
 
-module.exports = {
+export {
   throwBadRequest,
   throwBadRequestWithArray,
   throwForbidden,
