+++ marklogic-data-hub/src/main/resources/ml-modules/root/data-hub/5/impl/hub-utils.mjs
@@ -212,7 +212,7 @@ function requireFunction(modulePath, functionName) {
   return cachedLibraries[modulePath][functionName];
 }
 
-module.exports = {
+export {
   capitalize,
   deleteDocument,
   documentsToContentDescriptorArray,
