+++ marklogic-data-hub/src/main/resources/ml-modules/root/data-hub/5/impl/hub-utils/invoke-single-delete.mjs
@@ -17,14 +17,8 @@
 
 declareUpdate();
 
-var docUri,
-    content,
-    permissions,
-    collections;
+const docUri = external.docUri;
 
-xdmp.documentInsert(docUri, content, {permissions: permissions, collections: collections });
-let writeInfo = {
-    transaction: xdmp.transaction(),
-    dateTime: fn.currentDateTime()
-};
-writeInfo;
+if (fn.docAvailable(docUri)) {
+    xdmp.documentDelete(docUri);
+}
