+++ marklogic-data-hub/src/main/resources/ml-modules/root/data-hub/5/impl/security.mjs
@@ -52,6 +52,6 @@ function getAuthorities() {
   return authorities;
 }
 
-module.exports = {
+export {
   getAuthorities
 };
